+++ depthai_ros_driver/include/depthai_ros_driver/dai_nodes/base_node.hpp
@@ -36,6 +36,9 @@ class BaseNode {
     std::string getName() {
         return baseDAINodeName;
     };
+    std::string getTFPrefix(const std::string& frameName = "") {
+        return std::string(getROSNode()->get_name()) + "_" + frameName;
+    }
 
    private:
     rclcpp::Node* baseNode;