+++ depthai_ros_driver/src/dai_nodes/sensors/rgb.cpp
@@ -21,7 +21,7 @@ RGB::RGB(const std::string& daiNodeName,
     ph->declareParams(node, colorCamNode, socket, sensor, publish);
     setXinXout(pipeline);
     RCLCPP_DEBUG(node->get_logger(), "Node %s created", daiNodeName.c_str());
-};
+}
 void RGB::setNames() {
     ispQName = getName() + "_isp";
     previewQName = getName() + "_preview";
@@ -33,60 +33,75 @@ void RGB::setXinXout(std::shared_ptr<dai::Pipeline> pipeline) {
         xoutColor = pipeline->create<dai::node::XLinkOut>();
         xoutColor->setStreamName(ispQName);
         if(ph->getParam<bool>(getROSNode(), "i_low_bandwidth")) {
-            RCLCPP_INFO(getROSNode()->get_logger(), "POE");
-
-            videoEnc = pipeline->create<dai::node::VideoEncoder>();
-            videoEnc->setQuality(ph->getParam<int>(getROSNode(), "i_low_bandwidth_quality"));
-            videoEnc->setProfile(dai::VideoEncoderProperties::Profile::MJPEG);
+            videoEnc = sensor_helpers::createEncoder(pipeline, ph->getParam<int>(getROSNode(), "i_low_bandwidth_quality"));
             colorCamNode->video.link(videoEnc->input);
             videoEnc->bitstream.link(xoutColor->input);
         } else {
             colorCamNode->isp.link(xoutColor->input);
         }
-        if(ph->getParam<bool>(getROSNode(), "i_enable_preview")) {
+    }
+    if(ph->getParam<bool>(getROSNode(), "i_enable_preview")) {
             xoutPreview = pipeline->create<dai::node::XLinkOut>();
             xoutPreview->setStreamName(previewQName);
             xoutPreview->input.setQueueSize(2);
             xoutPreview->input.setBlocking(false);
             colorCamNode->preview.link(xoutPreview->input);
         }
-    }
     xinControl = pipeline->create<dai::node::XLinkIn>();
     xinControl->setStreamName(controlQName);
     xinControl->out.link(colorCamNode->inputControl);
 }
 
 void RGB::setupQueues(std::shared_ptr<dai::Device> device) {
-    auto calibHandler = device->readCalibration();
     if(ph->getParam<bool>(getROSNode(), "i_publish_topic")) {
-        auto tfPrefix = std::string(getROSNode()->get_name()) + "_" + getName();
+        auto tfPrefix = getTFPrefix(getName());
+        infoManager = std::make_shared<camera_info_manager::CameraInfoManager>(
+            getROSNode()->create_sub_node(std::string(getROSNode()->get_name()) + "/" + getName()).get(), "/" + getName());
         imageConverter = std::make_unique<dai::ros::ImageConverter>(tfPrefix + "_camera_optical_frame", false);
-        colorQ = device->getOutputQueue(ispQName, ph->getParam<int>(getROSNode(), "i_max_q_size"), false);
-        colorQ->addCallback(std::bind(&RGB::colorQCB, this, std::placeholders::_1, std::placeholders::_2));
+        if(ph->getParam<std::string>(getROSNode(), "i_calibration_file").empty()) {
+            infoManager->setCameraInfo(sensor_helpers::getCalibInfo(getROSNode()->get_logger(),
+                                                                    *imageConverter,
+                                                                    device,
+                                                                    static_cast<dai::CameraBoardSocket>(ph->getParam<int>(getROSNode(), "i_board_socket_id")),
+                                                                    ph->getParam<int>(getROSNode(), "i_width"),
+                                                                    ph->getParam<int>(getROSNode(), "i_height")));
+        } else {
+            infoManager->loadCameraInfo(ph->getParam<std::string>(getROSNode(), "i_calibration_file"));
+        }
         rgbPub = image_transport::create_camera_publisher(getROSNode(), "~/" + getName() + "/image_raw");
-
-        if(ph->getParam<bool>(getROSNode(), "i_enable_preview")) {
-            previewQ = device->getOutputQueue(previewQName, ph->getParam<int>(getROSNode(), "i_max_q_size"), false);
-            previewQ->addCallback(std::bind(&RGB::colorQCB, this, std::placeholders::_1, std::placeholders::_2));
-            previewPub = image_transport::create_camera_publisher(getROSNode(), "~/" + getName() + "/preview/image_raw");
-            try {
-                previewInfo = imageConverter->calibrationToCameraInfo(calibHandler,
-                                                                      static_cast<dai::CameraBoardSocket>(ph->getParam<int>(getROSNode(), "i_board_socket_id")),
-                                                                      ph->getParam<int>(getROSNode(), "i_preview_size"),
-                                                                      ph->getParam<int>(getROSNode(), "i_preview_size"));
-            } catch(std::runtime_error& e) {
-                RCLCPP_ERROR(getROSNode()->get_logger(), "No calibration! Publishing empty camera_info.");
-            }
-        };
-        try {
-            rgbInfo = imageConverter->calibrationToCameraInfo(calibHandler,
-                                                              static_cast<dai::CameraBoardSocket>(ph->getParam<int>(getROSNode(), "i_board_socket_id")),
-                                                              ph->getParam<int>(getROSNode(), "i_width"),
-                                                              ph->getParam<int>(getROSNode(), "i_height"));
-        } catch(std::runtime_error& e) {
-            RCLCPP_ERROR(getROSNode()->get_logger(), "No calibration! Publishing empty camera_info.");
+        colorQ = device->getOutputQueue(ispQName, ph->getParam<int>(getROSNode(), "i_max_q_size"), false);
+        if(ph->getParam<bool>(getROSNode(), "i_low_bandwidth")) {
+            colorQ->addCallback(std::bind(sensor_helpers::compressedImgCB,
+                                          std::placeholders::_1,
+                                          std::placeholders::_2,
+                                          *imageConverter,
+                                          rgbPub,
+                                          infoManager,
+                                          dai::RawImgFrame::Type::BGR888i));
+        } else {
+            colorQ->addCallback(std::bind(sensor_helpers::imgCB, std::placeholders::_1, std::placeholders::_2, *imageConverter, rgbPub, infoManager));
         }
     }
+    if(ph->getParam<bool>(getROSNode(), "i_enable_preview")) {
+        previewQ = device->getOutputQueue(previewQName, ph->getParam<int>(getROSNode(), "i_max_q_size"), false);
+        previewPub = image_transport::create_camera_publisher(getROSNode(), "~/" + getName() + "/preview/image_raw");
+        previewInfoManager = std::make_shared<camera_info_manager::CameraInfoManager>(
+            getROSNode()->create_sub_node(std::string(getROSNode()->get_name()) + "/" + previewQName).get(), previewQName);
+        auto tfPrefix = getTFPrefix(getName());
+        imageConverter = std::make_unique<dai::ros::ImageConverter>(tfPrefix + "_camera_optical_frame", false);
+        if(ph->getParam<std::string>(getROSNode(), "i_calibration_file").empty()) {
+            previewInfoManager->setCameraInfo(
+                sensor_helpers::getCalibInfo(getROSNode()->get_logger(),
+                                             *imageConverter,
+                                             device,
+                                             static_cast<dai::CameraBoardSocket>(ph->getParam<int>(getROSNode(), "i_board_socket_id")),
+                                             ph->getParam<int>(getROSNode(), "i_preview_size"),
+                                             ph->getParam<int>(getROSNode(), "i_preview_size")));
+        } else {
+            infoManager->loadCameraInfo(ph->getParam<std::string>(getROSNode(), "i_calibration_file"));
+        }
+        previewQ->addCallback(std::bind(sensor_helpers::imgCB, std::placeholders::_1, std::placeholders::_2, *imageConverter, previewPub, previewInfoManager));
+    };
     controlQ = device->getInputQueue(controlQName);
 }
 
@@ -98,26 +113,6 @@ void RGB::closeQueues() {
         }
     }
     controlQ->close();
-}
-
-void RGB::colorQCB(const std::string& name, const std::shared_ptr<dai::ADatatype>& data) {
-    auto img = std::dynamic_pointer_cast<dai::ImgFrame>(data);
-    std::deque<sensor_msgs::msg::Image> deq;
-    if(ph->getParam<bool>(getROSNode(), "i_low_bandwidth"))
-        imageConverter->toRosMsgFromBitStream(img, deq, dai::RawImgFrame::Type::BGR888i, rgbInfo);
-    else
-        imageConverter->toRosMsg(img, deq);
-    while(deq.size() > 0) {
-        auto currMsg = deq.front();
-        if(name == ispQName) {
-            rgbInfo.header = currMsg.header;
-            rgbPub.publish(currMsg, rgbInfo);
-        } else {
-            previewInfo.header = currMsg.header;
-            previewPub.publish(currMsg, previewInfo);
-        }
-        deq.pop_front();
-    }
 }
 
 void RGB::link(const dai::Node::Input& in, int linkType) {